Closes #37: Sort popular content by popularity descending

+++ app/src/main/java/com/vignesh/moviebucket/data/source/local/MovieLocalDataSource.kt
@@ -155,7 +155,7 @@ class MovieLocalDataSource(
                 getString(TITLE),
                 getString(RELEASE_DATE),
                 getInt(RUNTIME),
-                getString(POPULARITY),
+                getString(POPULARITY).toDouble(),
                 getString(RATING),
                 getString(GENRE),
                 getString(OVERVIEW),
